% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{appendix_table}
\alias{appendix_table}
\title{Create a table to go in the report appendix}
\usage{
appendix_table(interview_data, gear, variable)
}
\description{
Create a table to go in the report appendix
}
