% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-bootstrap.R
\name{bootstrap_harvest}
\alias{bootstrap_harvest}
\title{Perform Non-Parametric Bootstrapping On Harvest Estimates}
\usage{
bootstrap_harvest(
  interview_data,
  effort_info,
  gear,
  n_boot = 1000,
  source_weights = NULL,
  stratify_interviews = TRUE
)
}
\description{
Perform Non-Parametric Bootstrapping On Harvest Estimates
}
