% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_datetime}
\alias{combine_datetime}
\title{Convert date and time variables into a datetime variable}
\usage{
combine_datetime(dates, times)
}
\description{
Convert date and time variables into a datetime variable
}
