% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{create_harvest_combos}
\alias{create_harvest_combos}
\title{Create combinations of interview data sources to use for harvest sensitivity analysis}
\usage{
create_harvest_combos(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Create combinations of interview data sources to use for harvest sensitivity analysis
}
\details{
Based on the data sources that conducted interviews, constructs a set of combinations
of data scenarios where some sources are discarded. Each scenario (beyond the original set)
includes just one data source, unlike \code{\link[=create_interview_combos]{create_interview_combos()}}.
}
