% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-tool-data.R
\name{data_tool}
\alias{data_tool}
\title{Interactively prepare interview and flight data}
\usage{
data_tool()
}
\description{
An RStudio add-in that guides the user through preparing the raw
data files to be used by other components of 'KuskoHarvEst'
}
