% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report.R
\name{effort_bullets}
\alias{effort_bullets}
\title{Summarize Effort Estimates Bullet List}
\usage{
effort_bullets(do_drift, do_set)
}
\arguments{
\item{do_drift}{Logical; should drift net effort estimates (and corresponding estimator info) be returned?}

\item{do_set}{Logical; should set net effort estimates be returned?}
}
\description{
Depending on which gears to report,
builds the appropriate summary bullet list.
}
