% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-trips-plot.R
\name{effort_plot}
\alias{effort_plot}
\title{Create a plot summarizing effort estimation}
\usage{
effort_plot(flight_data, effort_info, gear)
}
\description{
Create a plot summarizing effort estimation
}
