% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{effort_sensitivity_table}
\alias{effort_sensitivity_table}
\title{Build a table to report results of effort sensitivity analyses}
\usage{
effort_sensitivity_table(effort_scenarios, flight_data, combos)
}
\description{
Build a table to report results of effort sensitivity analyses
}
