% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-catch-per-trip.R
\name{estimate_catch_per_trip}
\alias{estimate_catch_per_trip}
\title{Estimate expected catch for the average trip}
\usage{
estimate_catch_per_trip(
  interview_data,
  gear,
  randomize = FALSE,
  central_fn = getOption("central_fn")
)
}
\description{
Estimate expected catch for the average trip
}
