% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-harvest.R
\name{estimate_harvest}
\alias{estimate_harvest}
\title{Estimate harvest}
\usage{
estimate_harvest(
  interview_data,
  effort_est,
  gear,
  include_whitefishes = FALSE,
  central_fn = mean
)
}
\description{
Estimate harvest
}
