% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-harvest-all.R
\name{estimate_harvest_all}
\alias{estimate_harvest_all}
\title{Estimate Harvest for All Strata}
\usage{
estimate_harvest_all(
  interview_data,
  effort_info,
  gear,
  stratify_interviews = TRUE,
  pooling_threshold = 10
)
}
\description{
Estimate Harvest for All Strata
}
