% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{flight_data_table}
\alias{flight_data_table}
\title{Create a table to report the flights and the counts that were made on each}
\usage{
flight_data_table(flight_data)
}
\description{
Create a table to report the flights and the counts that were made on each
}
