% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{get_flight_strata}
\alias{get_flight_strata}
\title{Find Which Strata are in Flight Data}
\usage{
get_flight_strata(flight_data)
}
\arguments{
\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}
}
\description{
Extracts the stratum names found in the flight data
and returns an error if any are present that are not accepted.
}
