% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report.R
\name{harvest_bullets}
\alias{harvest_bullets}
\title{Summarize Harvest Estimates Bullet List}
\usage{
harvest_bullets(include_set_summary, nonsalmon = FALSE)
}
\arguments{
\item{include_set_summary}{Logical; do you wish to return a bullet point reporting the total estimated harvest by set nets as well as the species composition?}

\item{nonsalmon}{Logical; should estimates be returned for whitefish and sheefish rather than for Chinook, chum, and sockeye salmon?}
}
\description{
Wrapper around \code{\link[=report]{report()}} to cleanly return harvest estimates
as a markdown list. Automatically selects species to report.
}
