% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{has_net_length}
\alias{has_net_length}
\title{Check if the net length was recorded}
\usage{
has_net_length(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Check if the net length was recorded
}
