% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{has_soak}
\alias{has_soak}
\title{Check if soak duration was recorded}
\usage{
has_soak(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Check if soak duration was recorded
}
