% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{has_trip_times}
\alias{has_trip_times}
\title{Check if the trip times were recorded}
\usage{
has_trip_times(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Check if the trip times were recorded
}
