% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{install_TeX}
\alias{install_TeX}
\title{Install All Necessary TeX Dependencies}
\usage{
install_TeX(force = TRUE)
}
\arguments{
\item{force}{Logical; If \code{TRUE}, will install TinyTeX even if a version is already installed.
TeX packages are installed regardless.}
}
\description{
Including tinytex and all needed TeX packages
}
