% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{interview_data_table}
\alias{interview_data_table}
\title{Create a table to report the number of interviews by data source}
\usage{
interview_data_table(interview_data)
}
\description{
Create a table to report the number of interviews by data source
}
