% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_catch_per_trip_outlier}
\alias{is_catch_per_trip_outlier}
\title{Determine if an interview is an outlier in terms of catch per trip}
\usage{
is_catch_per_trip_outlier(
  interview_data,
  catch_per_trip_cut = getOption("catch_per_trip_cut")
)
}
\description{
Determine if an interview is an outlier in terms of catch per trip
}
