% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_complete_trip}
\alias{is_complete_trip}
\title{Check if interview is for a completed trip}
\usage{
is_complete_trip(interview_data)
}
\description{
Check if interview is for a completed trip
}
\details{
Data need special care if from an incomplete trip.
E.g., soak time should not be used in calculation of the average across trips,
and if the soak time is excessively short at the time of the interview relative to completed
trips, it shouldn't be used to inform the average catch rate.
}
