% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_normal_net}
\alias{is_normal_net}
\title{Check if the net characteristics are as expected}
\usage{
is_normal_net(
  interview_data,
  net_length_cut = KuskoHarvEst_opts("net_length_cut")
)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{net_length_cut}{Numeric; maximum net length (in feet) allowed, records with net length longer
than this will have their net length deemed unsuitable for any purpose}
}
\description{
Check if the net characteristics are as expected
}
\details{
Nets that are extremely long are probably mis-recorded and should not be included.
}
