% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_possible_soak}
\alias{is_possible_soak}
\title{Check for impossible soak times}
\usage{
is_possible_soak(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Check for impossible soak times
}
\details{
If the soak duration is longer than the trip duration, this is impossible and an error.
}
