% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_short_incomplete_soak}
\alias{is_short_incomplete_soak}
\title{Check incomplete trips for very short soak times}
\usage{
is_short_incomplete_soak(interview_data)
}
\description{
Check incomplete trips for very short soak times
}
\details{
For incomplete trips, if the interview was conducted early into an incomplete
trip, its catch rate information is not likely representative of that from completed trips.
This function looks for incomplete interviews that have soak time shorter than the shortest recorded
soak time among completed trips for that gear type. Records with a \code{TRUE} value returned should
not return
}
