% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{johnson_summary_table}
\alias{johnson_summary_table}
\title{Create a table to summarize catch rates and species composition relative to Johnson River}
\usage{
johnson_summary_table()
}
\description{
Create a table to summarize catch rates and species composition relative to Johnson River
}
