% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{KuskoHarvEst_opts}
\alias{KuskoHarvEst_opts}
\title{Set or Get KuskoHarvEst Options}
\usage{
KuskoHarvEst_opts(...)
}
\arguments{
\item{...}{Option names to retrieve options values or \code{key = value} pairs
to set options. Accepted settings include:
\itemize{
\item \code{soak_sd_cut} (default: 3)
\item \code{net_length_cut} (default: 350)
\item \code{catch_per_trip_cut} (default: 0.05)
\item \code{central_fn} (default: mean)
\item \code{pooling_threshold} (default: 10)
\item \code{interview_data} (default: \code{NULL})
\item \code{flight_data} (default: \code{NULL})
\item \code{boot_out} (default: \code{NULL})
}}
}
\description{
Set or Get KuskoHarvEst Options
}
\note{
The options \code{interview_data}, \code{flight_data}, and \code{boot_out}
are not currently used in any function. I'm leaving them here
as placeholders for the future.
}
