% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{KuskoHarvEst_skeleton}
\alias{KuskoHarvEst_skeleton}
\title{Create a project directory to use with 'KuskoHarvEst'}
\usage{
KuskoHarvEst_skeleton(path)
}
\description{
Create a project directory to use with 'KuskoHarvEst'
}
