% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{link_to_doc}
\alias{link_to_doc}
\title{Create a markdown link to a local documentation file}
\usage{
link_to_doc(doc, text = "here")
}
\arguments{
\item{doc}{Character; file path to the documentation file in question}

\item{text}{Character; the text to display as the clickable link}
}
\description{
Create a markdown link to a local documentation file
}
