% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_appendix_tables}
\alias{make_appendix_tables}
\title{Create multiple tables for the appendix}
\usage{
make_appendix_tables(
  interview_data,
  gear,
  include_rates = TRUE,
  split_chum_sockeye = TRUE,
  extra = c("trip_duration", "soak_duration"),
  nonsalmon = FALSE
)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{gear}{Character; which gear type to use? Only two options are accepted:
\itemize{
\item \code{gear = "drift"} for drift nets
\item \code{gear = "set"} for set nets
}}

\item{include_rates}{Logical; do you wish to include tables summarizing catch rates for the available species
(expressed as catch/150ft of net/1hr, defaults to \code{TRUE})?}

\item{split_chum_sockeye}{Logical; if both chum and sockeye salmon data are available, do you wish to
present their summaries disaggregated by species (\code{TRUE}, default), or aggregated together as chum+sockeye (\code{FALSE})?}

\item{extra}{Character; Any extra variables to include summary tables for. Set to \code{NULL} to exclude any extras.
Accepted options include:
* \code{"trip_duration"} (included by default)
* \code{"soak_duration"} (included by default)
* \code{"trip_start"}
* \code{"trip_end"}
* \code{"net_length"}}

\item{nonsalmon}{Logical; should estimates be returned for whitefish and sheefish rather than for Chinook, chum, and sockeye salmon?}
}
\description{
Create multiple tables for the appendix
}
\note{
Unlike the other functions that make tables (e.g., \code{\link[=make_strata_summary_table]{make_strata_summary_table()}}),
this function must be used with the chunk option \code{results = "asis"} to render properly.
Additionally, if Chinook salmon are available and \code{nonsalmon = FALSE} is set, a table showing
the percent composition of Chinook salmon among all available salmon species will be included.
}
