% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-trips-plot.R
\name{make_effort_plot}
\alias{make_effort_plot}
\title{Create a plot summarizing effort estimation}
\usage{
make_effort_plot(flight_data, effort_info, trips_only = FALSE)
}
\arguments{
\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}

\item{effort_info}{List storing the output of \code{\link[=estimate_effort]{estimate_effort()}}}

\item{trips_only}{Logical; should figure only show trip/flight times, or should
effort estimation summaries be displayed at the bottom of the figure?}
}
\description{
Create a plot summarizing effort estimation
}
