% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{make_effort_sensitivity_table}
\alias{make_effort_sensitivity_table}
\title{Make a table to report results of effort sensitivity analyses}
\usage{
make_effort_sensitivity_table(effort_scenarios, flight_data, combos)
}
\description{
Make a table to report results of effort sensitivity analyses
}
