% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{make_flight_combos}
\alias{make_flight_combos}
\title{Make combinations of flights to leave out for effort sensitivity analysis}
\usage{
make_flight_combos(flight_data)
}
\description{
Make combinations of flights to leave out for effort sensitivity analysis
}
