% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_flight_data_table}
\alias{make_flight_data_table}
\title{Create a table to report the flights and the counts that were made on each}
\usage{
make_flight_data_table(flight_data)
}
\arguments{
\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}
}
\description{
Create a table to report the flights and the counts that were made on each
}
