% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_goals_summary_table}
\alias{make_goals_summary_table}
\title{Create a table displaying reported harvest goal attainment}
\usage{
make_goals_summary_table(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Create a table displaying reported harvest goal attainment
}
