% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{make_harvest_sensitivity_tables}
\alias{make_harvest_sensitivity_tables}
\title{Make a table to report multiple tables from harvest sensitivity analyses}
\usage{
make_harvest_sensitivity_tables(harvest_scenarios, combos)
}
\arguments{
\item{harvest_scenarios}{List storing the bootstrapped harvest output from each data scenario}

\item{combos}{The output of \code{\link[=make_harvest_combos]{make_harvest_combos()}}}
}
\description{
Make a table to report multiple tables from harvest sensitivity analyses
}
\note{
Unlike most other functions that make tables (e.g., \code{\link[=make_strata_summary_table]{make_strata_summary_table()}}),
this function must be used with the chunk option \code{results = "asis"} to render properly.
}
