% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-histograms.R
\name{make_histogram}
\alias{make_histogram}
\title{Make a single histogram showing distribution of a quantity across interviews}
\usage{
make_histogram(interview_data, gear, variable, n_bins = 10)
}
\description{
Make a single histogram showing distribution of a quantity across interviews
}
