% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-histograms.R
\name{make_histograms}
\alias{make_histograms}
\title{Make a multi-panel figure showing multiple histograms}
\usage{
make_histograms(interview_data, gear, variables, n_bins = 10)
}
\description{
Make a multi-panel figure showing multiple histograms
}
