% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_johnson_summary_table}
\alias{make_johnson_summary_table}
\title{Create a table to summarize catch rates and species composition relative to Johnson River}
\usage{
make_johnson_summary_table()
}
\description{
Create a table to summarize catch rates and species composition relative to Johnson River
}
\details{
Requires that an object named \code{boot_out} to be in existence
}
