% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_strata_summary_table}
\alias{make_strata_summary_table}
\title{Create a table to summarize information spatially}
\usage{
make_strata_summary_table(interview_data, gear, nonsalmon = FALSE)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{gear}{Character; which gear type to use? Only two options are accepted:
\itemize{
\item \code{gear = "drift"} for drift nets
\item \code{gear = "set"} for set nets
}}

\item{nonsalmon}{Logical; should the table display estimates for non-salmon instead of salmon species?}
}
\description{
Create a table to summarize information spatially
}
