% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-effort.R
\name{N_estimator}
\alias{N_estimator}
\title{Estimate the Number of Trips that Occurred in a Fishing Day}
\usage{
N_estimator(effort_data)
}
\arguments{
\item{effort_data}{A \code{\link[base:list]{list}} storing the number of trips reported being active on individual, consecutive, and not any flights.
Constructed via \code{\link[=tally_effort_data]{tally_effort_data()}}.}
}
\value{
A numeric vector with named elements for:
\itemize{
\item \code{pi_hat} (or \code{piF_hat}, where \code{F} is a flight number) -- the estimated proportion (probability) of all trips counted on each flight
\item \code{psi_hat} -- the estimated proportion (probability) of all trips that were interviewed
\item \code{N_hat} -- the estimated total number of trips that occurred that day
}
}
\description{
Based on the number of interviewed trips that
reported being active during each aerial survey
}
\details{
Performs the calculations as specified by estimators derived in the appendix of an in-prep manuscript.
Suffice it to say that the estimator combines the interview data and flight data to estimate the fraction
of all trips that were interviewed and counted by aerial survey.
These form the basis for expanding the partial counts to the whole. In the case of multiple flights,
the estimator accounts for trips that would have been counted on more than one flight.
}
