% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-rmd-output-types.R
\name{pdf_report}
\alias{pdf_report}
\title{Rmarkdown output format for PDF reports}
\usage{
pdf_report(...)
}
\arguments{
\item{...}{Optional arguments supplied to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}}
}
\description{
A custom output type for PDFs produced by 'KuskoHarvEst'
}
\details{
The \code{template} argument and \code{latex_engine} arguments are defined internally and cannot be
supplied to \code{...}. \code{latex_engine = "pdflatex"} is used.
}
