% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-flights.R
\name{prepare_flights}
\alias{prepare_flights}
\title{Prepare flight data file}
\usage{
prepare_flights(input_file)
}
\arguments{
\item{input_file}{Character; the name of the flight data file}
}
\description{
Reads in and formats the data file containing flight counts of active fishing trips
into a standardized format to be used by other functions
}
