% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-interviews.R
\name{prepare_interviews}
\alias{prepare_interviews}
\title{Format interview data for analysis}
\usage{
prepare_interviews(input_file, include_village = FALSE, include_goals = FALSE)
}
\description{
Format interview data for analysis
}
