% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-interviews-one.R
\name{prepare_interviews_one}
\alias{prepare_interviews_one}
\title{Prepare one interview data file}
\usage{
prepare_interviews_one(
  input_file,
  include_salmon,
  include_nonsalmon,
  include_village,
  include_goals
)
}
\arguments{
\item{input_file}{Character; name of a single file that contains interview data}

\item{include_salmon}{Character; accepted options are \code{"all"}, \code{"none"}, or any combination of acceptable salmon species names}

\item{include_nonsalmon}{Character; accepted options are \code{"all"}, \code{"none"}, or any combination of acceptable nonsalmon species names}

\item{include_village}{Logical; should the village of the fisher be included in the output?}

\item{include_goals}{Logical; should the fisher's reported progress towards meeting their season-wide harvest goals be returned?}
}
\description{
Reads in and formats one data file containing interview data
into a standardized format to be used by other functions
}
\note{
This function should always be called by calling the wrapper \code{\link[=prepare_interviews]{prepare_interviews()}}, not directly.
}
