% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-randomize.R
\name{randomize_data}
\alias{randomize_data}
\title{Sample with replacement from a data set}
\usage{
randomize_data(interview_data)
}
\arguments{
\item{interview_data}{data frame generated with \code{prepare_interviews()}
See details.}
}
\description{
Sample with replacement from a data set
}
