% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report.R
\name{report}
\alias{report}
\title{Return summarized harvest estimates}
\usage{
report(
  spp = "total",
  gear = "total",
  stratum = "total",
  CI = TRUE,
  conf_level = 0.95,
  digits = -1,
  return_numeric = FALSE,
  boot_out_use = NULL
)
}
\description{
Return summarized harvest estimates
}
