% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-rmd-output-types.R
\name{resource_path}
\alias{resource_path}
\title{Create path to a package resource file}
\usage{
resource_path(file)
}
\arguments{
\item{file}{Character; a file name or file path within the resources folder to point to}
}
\description{
Constructs a complete file path to a package resource file located in
\code{rstudio/templates/project/resources} within the 'KuskoHarvEst' library
}
