% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-tool-rmd.R
\name{rmd_tool}
\alias{rmd_tool}
\title{Interactively construct PDF reports}
\usage{
rmd_tool()
}
\description{
An RStudio add-in that guides the user through constructing and rendering
the Rmarkdown source scripts for in-season estimate and sensitivity analysis reports
}
