% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-histograms.R
\name{select_histogram_variables}
\alias{select_histogram_variables}
\title{Automate selection of histogram variables}
\usage{
select_histogram_variables(
  interview_data,
  split_chum_sockeye = TRUE,
  max_vars = 6,
  min_noncatch = 2,
  noncatch_vars = c("trip_duration", "soak_duration", "trip_start", "trip_end")
)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{max_vars}{Numeric; the maximum number of variables to return, defaults to 6 for a 2x3 multi-panel plot.}

\item{min_noncatch}{Numeric; the minimum number of noncatch variables to return, defaults to 2.}

\item{noncatch_vars}{Character; vector of variable names to show in addition to the catch variables.
As many valid options will be returned (in order as specified here) as room allows by \code{max_vars}.
Defaults to \code{c("trip_duration", "soak_duration", "trip_start", "trip_end")}.}
}
\description{
Automate selection of histogram variables
}
