% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{select_species}
\alias{select_species}
\title{Select Species after Data Processing}
\usage{
select_species(interview_data, knitr_params)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{knitr_params}{List; must contain the names of the desired/undesired
species as named elements with logical (\code{TRUE/FALSE}) values.
Generally provided via the \code{params} argument to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\description{
Given a data frame of interview data,
return the entire data frame
but including only those selected in
a knitr \code{params} list.
}
