% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{species_in_data}
\alias{species_in_data}
\title{Find Which Species are in Data Set}
\usage{
species_in_data(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Given a data frame of interview data,
return the species names separated into
salmon and non-salmon categories
}
