% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitable-for.R
\name{suitable_for}
\alias{suitable_for}
\title{Determine which records are suitable for particular tasks}
\usage{
suitable_for(interview_data, task)
}
\description{
A wrapper around a variety of "checking" functions, all of which are non-exported.
}
