% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-effort.R
\name{tally_effort_data}
\alias{tally_effort_data}
\title{Count Types of Interview/Flight Outcomes}
\usage{
tally_effort_data(interview_data, flight_data)
}
\arguments{
\item{interview_data}{A \code{\link[base:data.frame]{data.frame}} storing the data from each interview, constructed via \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}
}
\value{
A \code{\link[base:list]{list}} storing the tallies as elements:
\itemize{
\item \verb{$Y}: the number of trips interviewed
\item \verb{$X}: the number of trips counted on each flight (e.g., element \code{"X1"} stores the first flight total)
\item \verb{$XnY}: the number of interviewed trips active at various times. This includes elements storing totals for:
\itemize{
\item Individual flights (e.g., element \code{"X1&Y"} is the number of trips that were interviewed and active during flight 1)
\item Consecutive flights (e.g., element \code{"X1&X2&Y"}) is the number of trips that were interviewed and active during both flights 1 and 2)
\item Not any flights (element \code{"!Xany&Y"})
\item All flights (element \code{"Xall&Y"})
}
}
}
\description{
Tallies how many trips were counted on each flight and consecutive flights.
}
\note{
Any interviews that (a) are \code{FALSE} for \code{suit_effort} or that (b) do not have \code{gear == "drift"} will be discarded prior to tallies.
}
