% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{unique_start_dates}
\alias{unique_start_dates}
\title{Determine the unique start dates of all interviews}
\usage{
unique_start_dates(interview_data)
}
\description{
Determine the unique start dates of all interviews
}
