% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{unlist_dfs}
\alias{unlist_dfs}
\title{Combine a list of data frames}
\usage{
unlist_dfs(list)
}
\arguments{
\item{list}{List; each element is a data frame with identical headers}
}
\description{
Loops over elements of a list object, where each element is a data frame with identical headers
and applies \code{\link[base:cbind]{base::rbind()}} to them.
}
