% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-effort.R
\name{was_counted}
\alias{was_counted}
\title{ID Trips that Overlapped Aerial Survey(s)}
\usage{
was_counted(trip_times, flight_times)
}
\arguments{
\item{trip_times}{A \code{\link[base:data.frame]{data.frame}} storing the start and end times of individual trips (one row per interviewed trip).
Must have columns \code{trip_start} and \code{trip_end}.}

\item{flight_times}{A \code{\link[base:data.frame]{data.frame}} storing the start and end times of each flight (one row per flight).
Must have columns \code{start_time} and \code{end_time}}
}
\value{
A \code{\link[base:data.frame]{data.frame}} with columns labeled \code{"X1"}, \code{"X2"}, etc., where the number represents the flight number of the day.
Columns store logical (i.e., \code{TRUE} or \code{FALSE}) values that indicate whether each trip (individual rows) was active during each flight.
}
\description{
Determines whether each trip was active (based on reported trip times)
at the time of each flight, and therefore assumed to have been counted via aerial survey
}
\details{
5 seconds are added to the start time and subtracted from the end time before any calculation.
This prevents mistakenly identifying trips that technically overlap (i.e., start/end immediately as the flight was taking off/landing), but almost certainly weren't counted.
}
\note{
All records must be supplied in the \code{datetime} format, ready to be processed by, e.g.,
\code{\link[lubridate:duration]{lubridate::duration()}}, \code{\link[lubridate:interval]{lubridate::interval()}}, and \code{\link[lubridate:interval]{lubridate::int_overlaps()}}
}
