% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-rmd-build-Rmd.R
\name{build_estimate_report_Rmd}
\alias{build_estimate_report_Rmd}
\title{Automate creation of a Rmd source file for in-season reports}
\usage{
build_estimate_report_Rmd(
  do_drift,
  do_set,
  species = c("chinook", "chum", "sockeye"),
  include_johnson_table = TRUE,
  include_goal_table = FALSE,
  include_appendix = FALSE,
  save_bootstrap = TRUE,
  ...
)
}
\arguments{
\item{do_drift}{Logical; should harvest and effort estimates be produced for drift nets?}

\item{do_set}{Logical; should harvest and effort estimates be produced for set nets?}

\item{species}{Character; vector of accepted species (see \code{KuskoHarvEst:::species_names}).}

\item{include_johnson_table}{Logical; should the output of \code{\link[=make_johnson_summary_table]{make_johnson_summary_table()}} be included?}

\item{include_goal_table}{Logical; should the output of \code{\link[=make_goals_summary_table]{make_goals_summary_table()}} be included?}

\item{include_appendix}{Logical; should the many tables each produced by \code{\link[=make_appendix_table]{make_appendix_table()}} be included?}

\item{save_bootstrap}{Logical; should a code chunk be included that saves a file containing the bootstrap samples of harvest?}

\item{...}{Arguments to be passed to \code{\link[=build_yaml]{build_yaml()}}. Both \code{do_drift = TRUE/FALSE} and \code{do_set = TRUE/FALSE}
are required, the remainder of the arguments have defaults. See \code{\link[=build_yaml]{build_yaml()}}.}
}
\description{
Based on a set of supplied options, builds the Rmarkdown source file to build
an in-season report documenting key output from the sampling and estimation
for a single day of fishing
}
\details{
This function selects from the many Rmarkdown source scripts found in \verb{inst/rstudio/templates/project/resources/}
(subdirectories: \code{01-common} and \code{02-estimate-report} therein) to automate the construction of the report source code.
This was previously a major time bottle neck, since old code had to be repeatedly copied, pasted, and edited depending on
the features of the new case the code needed to be applied to.
}
