% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-Rmd.R
\name{build_estimate_report_Rmd}
\alias{build_estimate_report_Rmd}
\title{Automate creation of a Rmd source file for in-season reports}
\usage{
build_estimate_report_Rmd(
  draft = FALSE,
  do_setnets = TRUE,
  n_boot = 1000,
  include_johnson_table = TRUE,
  include_goal_table = FALSE,
  include_appendix = FALSE
)
}
\description{
Automate creation of a Rmd source file for in-season reports
}
