% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-rmd-build-Rmd.R
\name{build_sensitivity_report_Rmd}
\alias{build_sensitivity_report_Rmd}
\title{Automate creation of a Rmd source file for sensitivity analysis reports}
\usage{
build_sensitivity_report_Rmd(
  draft = FALSE,
  do_setnets = TRUE,
  n_boot = 1000,
  include_plots = FALSE
)
}
\description{
Automate creation of a Rmd source file for sensitivity analysis reports
}
