% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-Rmd.R
\name{build_yaml}
\alias{build_yaml}
\title{Automate creation of a YAML header based on pre-defined meta data}
\usage{
build_yaml(doc_type, draft)
}
\arguments{
\item{doc_type}{One of \code{"estimate_report"} or \code{"sensitivity_report"}}

\item{draft}{Logical. Should a draft watermark be printed in the rendered output?}
}
\description{
Automate creation of a YAML header based on pre-defined meta data
}
