% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{combine_datetime}
\alias{combine_datetime}
\title{Convert date and time variables into a datetime variable}
\usage{
combine_datetime(dates, times)
}
\arguments{
\item{dates}{Character; vector with elements in \code{M/D/YYYY} format}

\item{times}{Character; vector with elements in \code{H:MM} format}
}
\value{
A vector with of class \code{datetime}, with the time zone set to US/Alaska
}
\description{
Convert date and time variables into a datetime variable
}
