% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{create_flight_combos}
\alias{create_flight_combos}
\title{Create combinations of flights to use for effort sensitivity analysis}
\usage{
create_flight_combos(flight_data)
}
\arguments{
\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}
}
\description{
Create combinations of flights to use for effort sensitivity analysis
}
\details{
Based on the number of flights conducted, constructs a set of combinations
of data scenarios where some flights are discarded.
}
