% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{create_interview_combos}
\alias{create_interview_combos}
\title{Create combinations of interview data sources to use for effort sensitivity analysis}
\usage{
create_interview_combos(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Create combinations of interview data sources to use for effort sensitivity analysis
}
\details{
Based on the data sources that conducted interviews, constructs a set of combinations
of data scenarios where some sources are discarded. All combinations of keeping/leaving out each
data source are included, unlike \code{\link[=create_harvest_combos]{create_harvest_combos()}}.
}
