% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-effort.R
\name{estimate_effort}
\alias{estimate_effort}
\title{Estimate Effort}
\usage{
estimate_effort(interview_data, flight_data, gear, method = "dbl_exp")
}
\arguments{
\item{interview_data}{A \code{\link[base:data.frame]{data.frame}} storing the data from each interview, constructed via \code{\link[=prepare_interviews]{prepare_interviews()}}}

\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}

\item{method}{Character; which effort estimator should be applied? Only two options are accepted:
\itemize{
\item \code{method = "dbl_exp"} to perform corrections for trips counted on consecutive flights and trips not counted at all - generally used for drift nets
\item \code{method = "max_per_stratum"} to perform a simple calculation of the maximum number of trips counted in a given stratum - generally used for set nets. If this method is used, a further constraint is added to force the total effort estimate to be at least as large as the total number of interviews for that gear type.
}}
}
\description{
Estimates total effort (completed trips) that occurred in a day of fishing
for a given gear type
}
\details{
If \code{method = "dbl_exp"}, this function calls two other critical functions on the data: \code{\link[=tally_effort_data]{tally_effort_data()}} and \code{\link[=N_estimator]{N_estimator()}}.
Please see these functions for more information.
}
