% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-harvest.R
\name{estimate_harvest}
\alias{estimate_harvest}
\title{Estimate harvest for all strata and species}
\usage{
estimate_harvest(
  interview_data,
  effort_info,
  gear,
  randomize = FALSE,
  stratify_interviews = TRUE
)
}
\description{
Estimate harvest for all strata and species
}
