% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-get-use-strata.R
\name{get_use_strata}
\alias{get_use_strata}
\title{Create stratum pooling strategy for interviews}
\usage{
get_use_strata(interview_data, pooling_threshold = 10, gear = "drift")
}
\description{
Create stratum pooling strategy for interviews
}
