% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-utils.R
\name{install_TeX}
\alias{install_TeX}
\title{Install All Necessary TeX Dependencies}
\usage{
install_TeX(remove_existing = TRUE)
}
\arguments{
\item{remove_existing}{Logical; do you wish to delete the existing
distribution found on your computer before installing? Defaults to \code{TRUE}.}
}
\description{
Including tinytex and all needed TeX packages
}
