% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_normal_net}
\alias{is_normal_net}
\title{Check if the net characteristics are as expected}
\usage{
is_normal_net(interview_data, length_cut = 350)
}
\description{
Check if the net characteristics are as expected
}
\details{
Nets that are extremely long are probably mis-recorded and should not be included.
Could feasibly include a mesh cutoff here if there was interest.
}
