% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_possible_soak}
\alias{is_possible_soak}
\title{Check for impossible soak times}
\usage{
is_possible_soak(interview_data)
}
\description{
Check for impossible soak times
}
\details{
If the soak duration is longer than the trip duration, this is impossible and an error.
The user should fix this entry in the data set before proceeding.
}
