% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{is_possible_trip}
\alias{is_possible_trip}
\title{Check for impossible trip times}
\usage{
is_possible_trip(interview_data)
}
\description{
Check for impossible trip times
}
\details{
If the trip end time is before the trip start time, this is impossible and an error.
The user should fix this entry in the data set before proceeding.
}
