% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-trips-plot.R
\name{make_effort_plot}
\alias{make_effort_plot}
\title{Create a plot summarizing effort estimation}
\usage{
make_effort_plot(flight_data, effort_info, trips_only = FALSE)
}
\description{
Create a plot summarizing effort estimation
}
