% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{make_effort_sensitivity_table}
\alias{make_effort_sensitivity_table}
\title{Make a table to report results of effort sensitivity analyses}
\usage{
make_effort_sensitivity_table(effort_scenarios, flight_data, combos)
}
\arguments{
\item{effort_scenarios}{List; contains the output of executing \code{\link[=estimate_effort]{estimate_effort()}}
repeatedly, once per combination of flights or interview data sources. Each list element is the output from a different
data scenario.}

\item{flight_data}{A \code{\link[base:data.frame]{data.frame}} storing the times and counts on each flight, constructed via \code{\link[=prepare_flights]{prepare_flights()}}}

\item{combos}{The output of \code{\link[=make_flight_combos]{make_flight_combos()}} or \code{\link[=make_interview_combos]{make_interview_combos()}}}
}
\description{
Make a table to report results of effort sensitivity analyses
}
