% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_goals_summary_table}
\alias{make_goals_summary_table}
\title{Create a table displaying reported harvest goal attainment}
\usage{
make_goals_summary_table(interview_data)
}
\description{
Create a table displaying reported harvest goal attainment
}
