% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-sensitivity.R
\name{make_harvest_sensitivity_table}
\alias{make_harvest_sensitivity_table}
\title{Make a table to report results from harvest sensitivity analyses}
\usage{
make_harvest_sensitivity_table(harvest_scenarios, combos)
}
\arguments{
\item{harvest_scenarios}{List storing the bootstrapped harvest output from each data scenario}

\item{combos}{The output of \code{\link[=make_harvest_combos]{make_harvest_combos()}}}
}
\description{
Make a table to report results from harvest sensitivity analyses
}
