% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-histograms.R
\name{make_histogram_caption}
\alias{make_histogram_caption}
\title{Make a figure caption for multipanel histograms}
\usage{
make_histogram_caption(interview_data, gear)
}
\description{
Make a figure caption for multipanel histograms
}
