% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-out-report-tables.R
\name{make_strata_summary_table}
\alias{make_strata_summary_table}
\title{Create a table to summarize information spatially}
\usage{
make_strata_summary_table(interview_data, gear)
}
\description{
Create a table to summarize information spatially
}
