% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-interviews-one.R
\name{prepare_interviews_one}
\alias{prepare_interviews_one}
\title{Prepare one interview data file}
\usage{
prepare_interviews_one(
  input_file,
  include_village = FALSE,
  include_goals = FALSE
)
}
\arguments{
\item{input_file}{Character; the name of one file that contains interview data from a single source and fishing day}

\item{include_village}{Logical; should the village of the fisher be included in the output?}

\item{include_goals}{Logical; should the fisher's reported progress towards meeting their season-wide harvest goals be returned?}
}
\description{
Reads in and formats one data file containing interview data
into a standardized format to be used by other functions
}
