% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-interviews-one.R
\name{prepare_interviews_one}
\alias{prepare_interviews_one}
\title{Format interview data for analysis}
\usage{
prepare_interviews_one(
  input_file,
  include_village = FALSE,
  include_goals = FALSE
)
}
\description{
Format interview data for analysis
}
