% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-randomize.R
\name{randomize_data}
\alias{randomize_data}
\title{Sample with replacement from a data set}
\usage{
randomize_data(interview_data, source_weights = NULL)
}
\arguments{
\item{interview_data}{data frame generated with \code{prepare_interviews()}}

\item{source_weights}{named vector representing sampling weights for different interview data sources.
See details.}
}
\description{
Sample with replacement from a data set
}
\details{
Some details here about the sample weights.
}
