% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-randomize.R
\name{randomize_data}
\alias{randomize_data}
\title{Sample with replacement from a data set}
\usage{
randomize_data(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Samples with replacement from a data frame containing interview data
}
