% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-est-stratify-effort.R
\name{stratify_effort}
\alias{stratify_effort}
\title{Apportion total effort estimate to strata using a simple rule}
\usage{
stratify_effort(flight_data, gear, effort_est)
}
\description{
Apportion total effort estimate to strata using a simple rule
}
