% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data-checks.R
\name{unique_start_dates}
\alias{unique_start_dates}
\title{Determine the unique start dates of all interviews}
\usage{
unique_start_dates(interview_data)
}
\arguments{
\item{interview_data}{Data frame storing interview data; created using \code{\link[=prepare_interviews]{prepare_interviews()}}}
}
\description{
Determine the unique start dates of all interviews
}
